Implement novel loader filtering system
Allow a second argument to be passed to presets, containing a callback
which will run on each generated loader definition. In the same manner
as a WordPress filter function, this callback can pass through a given
definition unchanged, or else (based on its loader type) adjust the
loader settings as desired and return a modified object.