Set production manifests to share seeds per output folder
@goldenapples notes in #153 that generating a production manifest
automatically could result in manifests being overwritten if two configs
share an output folder but do not specifically instruct their manifests
to use a consistent seed object. This commit changes the generation of
the production manifest to include a seed object which is unique per
output.path directory. In this way (validated by the changes to the
test-build npm script) multi-config webpack setups with multiple builds
targeting the same directory will share a manifest.