Use data in path when inferring publicPath
In a multiple-config Webpack setup, builds to two different output
directories appear to both need a specified, unique publicPath in order
for those bundles to be accessible as expected within the dev server.

To ensure that default usage creates bundles which can be loaded
correctly in a multi-config DevServer environment, we adapt the value
of output.path (stripping out the filePath cwd base) for use as a
subdirectory within the in-memory virtual filesystem of the DevServer.