Fix handling of RTL files in the manifest
When using the WebpackRTLPlugin the manifest doesn't produce entries
for both the original and the RTL file. This results in the wrong file
being loaded. This is caused by the fact that both files are in the
same chunk.

There is an issue tracking this on the WebpackRTLPlugin:
https://github.com/romainberger/webpack-rtl-plugin/issues/14

This commit includes a workaround. By providing a `map` function to
the manifest plugin we can change the name when we're dealing with an
RTL file. That results in two entries in the eventual manifest.