Inject development-mode asset manifest even without publicPath
This resolves an issue where multi-config entries would only output
a manifest containing the entries from the first exported config.

The origin of this issue is that we cannot define a devServer on
every multi-config array, or DevServer v4 will error. But without
defining a devServer we did not usually have the information
needed to deduce a publicPath, and therefore did not insert the
manifest plugin in subsequent configs.

Remove test for not setting manifest plugin